requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-crispy-forms==1.12.0
Pillow==8.3.1
psycopg2-binary